Fix timestamp bug in file reporter
The file reporter writes the profile into a local file, current
timestamp is added in the file name, unfortunately that timestamp was
set at load time because of the default argument so each reports would
always go into the same file.

This small fix will ensure that different files will be created and the
file name will get proper timestamp.