Add synthetic frame for lxml schema init
Creating schema in lxml can be a performance issue if we always recreate
a schema instead of reusing it.

Unfortunately this code is cython and does not appear in the flame
graph, this change searches for `etree.XMLSchema(` in the source code
line to add a synthetic frame for it to appear in the graph.

Tested with a local script.