Add synthetic frame for `queue.get(block=True)`
Whenever some python code is waiting for an item to arrive in a queue,
the current flame grpah will only show the function where the queue.get
is called, the rest of the stack is not showing. This is unfortunate
because we would need to remap those stacks into WAITING state.

This change adds another synthetic frame for this specific frame when
the source code at the end of the sampled stack has `.get(block=True`
This looks like it could trigger a lot of false positives but it will
only happen when this is the leaf of the stack so we would need this
code to call some native code that does not appear in the stack but
where we spend a lot of time.