Add operation name frame in stacks with boto api calls
When using API calls from boto we often do not see the operation names
in the stack traces even though there is a dedicate function name. This
is because of the generic way that api calls are implemented in boto.
Unfortunately this is not convenient for profiling as we do not see in
flamegraphs which calls are using resources.

This change adds a specific check when we build the stacks to detect api
calls inside boto and if possible add a new frame with the operation
function name.

i.e. instead of
```
(...)
botocore.client:CloudWatch:_make_api_call
botocore.client:CloudWatch:_api_call
user_module:send_metric
```

we should end up with
```
(...)
botocore.client:CloudWatch:_make_api_call
botocore.client:CloudWatch:_api_call
botocore.client:CloudWatch:put_metric_data
user_module:send_metric
```

This new frame should help when investigating on flamegraphs but also
with automatic analysis.