Fixing bug in calculating active time since start (#40)
* Fixing bug in calculating active time since start

In case of lambda where there are pause times, it should be considered when calculating active time. Essentially total pause time should be removed from [end-start]. Earlier we were considering the last sample time to be the end of the profile but in some corner cases where there is huge pause after that we would include it in the calculation even though pause is after the "end" of the profile.

So updating the way we calculate "end" of the profile. For more detailed information, please have a look at the comments.

Also, added some debug logs.

* Fixing unit tests

* Updating comment and minor change in resume

* Fixing unit tests for latest changes