Send agent overhead values as int
We are currently sending the `memory_usage_mb` and `timeInMs` fields in
the agent overhead metadata, unfortunately the schema for those is INT
which means they are currently discarded when the profile is validated
at submission time. This does not prevent profiles from being accepted
but we lose the overhead data which means we will not show it in the UI.

This change converts them to int before we serialize the report.
It would be nice for the backend to accept decimals at least for the
memory since it is in MB and we may often be under 1MB. But at the
moment it is better to align to current schema.