Update lambda handler function extraction to work with python 3.9
When running on lambda functions, the CodeGuru Profiler module should be
loaded by lambda bootstrap framework instead of the customer's code.
Then we load the customer's handler function by calling the bootstarp
code directly.

We currently handle python 3.8, 3.7 and 3.6 (with some tweeks), however
for python 3.9 the lambda bootstrap module is now delegating to the
[runtime interface client](https://github.com/aws/aws-lambda-python-runtime-interface-client)
so our code must be changed to also work with this.

This changes moves the code about finding the proper function that
can load a handler function into a separate method. Unit tests are
updated to cope with this.